Add files via upload
This MATLAB code simulates the propagation and collision of photons in a 2D space.

First, it sets the number of photons to be simulated (Nphotons) and generates random numbers to simulate the location of the photons (xdata and ydata).

Then, it creates a histogram of the photons' location using the hist3 function. This histogram is created using two-dimensional bins that span a range of -5 to 5, with a bin size of 0.03. The xcut variable is created as a 1D array of intensities corresponding to the center row of the histogram.

Finally, the code generates three plots:

The first plot shows the photon locations as white dots on a black background.
The second plot shows a 3D surface plot of the Gaussian distribution of the photons based on the histogram.
The third plot shows the intensity distribution of the photons in the x-axis only.
Overall, this code is a simple simulation of photon propagation and collision, demonstrating the use of histograms and 2D surface plots in MATLAB.