Add files via upload
This code generates an animation of the electric field's vector field for a Laguerre-Gaussian beam.

First, the polarize function defines the polarization states of the electric field. Then, the parameters for the beam are defined: the beam waist w, the topological charge m, and the number of points N.

The electric field is then calculated on a meshgrid defined by x and y, and the intensity and phase of the field are plotted using plt.pcolormesh.

Next, a quiver plot is created to show the electric field's vector field. The update function updates the quiver plot for each time step of the animation, which is created using FuncAnimation and displayed as HTML using HTML(anim.to_jshtml()).

The animation loops through frames=np.linspace(0, 2*np.pi, 50) with an interval of 50 milliseconds between frames. In each frame, the polarization of the electric field is changed, and the new electric field's vector field is calculated and plotted using q.set_UVC.