Add files via upload
This code simulates a set of random photon locations in a two-dimensional plane and generates a histogram of their distribution.

First, the necessary libraries are imported, including numpy for mathematical operations, matplotlib for plotting, and scipy for statistical functions. Then, the number of photons to simulate is defined, and their x and y positions are generated using numpy's random.normal function.

A scatter plot of the photon locations is created using matplotlib, with a black background and white dots. The title and axis labels are added to the plot.

Next, a 2D histogram of the photon positions is created using numpy's histogram2d function. The bins for the histogram are defined using the numpy arange function, and the resulting histogram is plotted using matplotlib's pcolormesh function. The title and axis labels are added to the plot.

Overall, this code provides a simple example of how to simulate random data and create a histogram of their distribution in Python using common libraries.